requirements.txt
diff-match-patch==20121119
Django==1.10
django-adminactions==1.5
django-bower==5.2.0
django-classy-tags==0.8.0
django-crispy-forms==1.6.1
django-geoposition==0.3.0
django-grappelli==2.9.1
django-import-export==0.5.1
et-xmlfile==1.0.1
freeze==1.0.10
future==0.16.0
jdcal==1.3
jsonfield==2.0.1
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
pdfkit==0.6.1
Pillow==4.1.0
psycopg2==2.7.1
pytz==2017.2
PyYAML==3.12
six==1.10.0
tablib==0.11.4
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
xlwt-future==0.8.0
